clean up annotation constants
