+++ ui/shared/components/panel/search/constants.js
@@ -6,192 +6,16 @@ import { CreateLocusListButton } from 'shared/components/buttons/LocusListButton
 import { RadioGroup, AlignedBooleanCheckbox, Select, InlineToggle } from 'shared/components/form/Inputs'
 import { snakecaseToTitlecase, camelcaseToTitlecase } from 'shared/utils/stringUtils'
 import {
-  VEP_GROUP_NONSENSE,
-  VEP_GROUP_ESSENTIAL_SPLICE_SITE,
-  VEP_GROUP_EXTENDED_SPLICE_SITE,
-  VEP_GROUP_MISSENSE,
-  VEP_GROUP_FRAMESHIFT,
-  VEP_GROUP_INFRAME,
-  VEP_GROUP_SYNONYMOUS,
-  VEP_GROUP_OTHER,
-  VEP_GROUP_SV,
-  VEP_GROUP_SV_CONSEQUENCES,
-  GROUPED_VEP_CONSEQUENCES,
   LOCUS_LIST_ITEMS_FIELD,
   ORDERED_PREDICTOR_FIELDS,
-  SPLICE_AI_FIELD,
-  VEP_GROUP_SV_NEW,
   PANEL_APP_CONFIDENCE_LEVELS,
-  SCREEN_LABELS,
+  SPLICE_AI_FIELD,
   predictorColorRanges,
 } from 'shared/utils/constants'
 
 import LocusListItemsFilter from './LocusListItemsFilter'
 import PaMoiSelector from './PaMoiSelector'
 import PaLocusListSelector from './PaLocusListSelector'
-
-export const ANNOTATION_GROUPS = Object.entries(GROUPED_VEP_CONSEQUENCES).map(([name, options]) => ({
-  name, options, groupLabel: snakecaseToTitlecase(name),
-}))
-
-const SCREEN_GROUP = 'SCREEN'
-const SCREEN_VALUES = ['PLS', 'pELS', 'dELS', 'DNase-H3K4me3', 'CTCF-only', 'DNase-only', 'low-DNase']
-const UTR_ANNOTATOR_GROUP = 'UTRAnnotator'
-const UTR_ANNOTATOR_VALUES = [
-  'premature_start_codon_gain', 'premature_start_codon_loss', 'stop_codon_gain', 'stop_codon_loss', 'uORF_frameshift',
-]
-const MOTIF_GROUP = 'motif_feature'
-const MOTIF_VALUES = [
-  {
-    description: 'A feature ablation whereby the deleted region includes a transcription factor binding site',
-    text: 'TFBS ablation',
-    value: 'TFBS_ablation',
-    so: 'SO:0001895',
-  },
-  {
-    description: 'A feature amplification of a region containing a transcription factor binding site',
-    text: 'TFBS amplification',
-    value: 'TFBS_amplification',
-    so: 'SO:0001892',
-  },
-  {
-    description: 'In regulatory region annotated by Ensembl',
-    text: 'TF binding site variant',
-    value: 'TF_binding_site_variant',
-    so: 'SO:0001782',
-  },
-  {
-    description: 'A fusion impacting a transcription factor binding site',
-    text: 'TFBS fusion',
-    value: 'TFBS_fusion',
-  },
-  {
-    description: 'A translocation impacting a transcription factor binding site',
-    text: 'TFBS translocation',
-    value: 'TFBS_translocation',
-  },
-]
-const REGULATORY_GROUP = 'regulatory_feature'
-const REGULATORY_VALUES = [
-  {
-    description: 'A sequence variant located within a regulatory region',
-    text: 'Regulatory region variant',
-    value: 'regulatory_region_variant',
-    so: 'SO:0001566',
-  },
-  {
-    description: 'A feature ablation whereby the deleted region includes a regulatory region',
-    text: 'Regulatory region ablation',
-    value: 'regulatory_region_ablation',
-    so: 'SO:0001894',
-  },
-  {
-    description: 'A feature amplification of a region containing a regulatory region',
-    text: 'Regulatory region amplification',
-    value: 'regulatory_region_amplification',
-    so: 'SO:0001891',
-  },
-  {
-    description: 'A fusion impacting a regulatory region',
-    text: 'Regulatory region fusion',
-    value: 'regulatory_region_fusion',
-  },
-]
-ANNOTATION_GROUPS.push({
-  name: SCREEN_GROUP,
-  groupLabel: SCREEN_GROUP,
-  options: SCREEN_VALUES.map(value => ({
-    value,
-    text: SCREEN_LABELS[value] || value,
-    description: 'SCREEN: Search Candidate cis-Regulatory Elements by ENCODE. Registry of cCREs V3’',
-  })),
-}, {
-  name: UTR_ANNOTATOR_GROUP,
-  groupLabel: UTR_ANNOTATOR_GROUP,
-  options: UTR_ANNOTATOR_VALUES.map(value => ({
-    value: `5_prime_UTR_${value}_variant`,
-    text: snakecaseToTitlecase(value),
-  })),
-}, {
-  name: MOTIF_GROUP,
-  groupLabel: snakecaseToTitlecase(MOTIF_GROUP),
-  options: MOTIF_VALUES,
-}, {
-  name: REGULATORY_GROUP,
-  groupLabel: snakecaseToTitlecase(REGULATORY_GROUP),
-  options: REGULATORY_VALUES,
-})
-
-const ALL_IMPACT_GROUPS = [
-  VEP_GROUP_NONSENSE,
-  VEP_GROUP_ESSENTIAL_SPLICE_SITE,
-  VEP_GROUP_EXTENDED_SPLICE_SITE,
-  VEP_GROUP_MISSENSE,
-  VEP_GROUP_FRAMESHIFT,
-  VEP_GROUP_INFRAME,
-  VEP_GROUP_SYNONYMOUS,
-  VEP_GROUP_OTHER,
-  VEP_GROUP_SV,
-  VEP_GROUP_SV_CONSEQUENCES,
-]
-const HIGH_IMPACT_GROUPS = [
-  VEP_GROUP_NONSENSE,
-  VEP_GROUP_ESSENTIAL_SPLICE_SITE,
-  VEP_GROUP_FRAMESHIFT,
-]
-export const ANNOTATION_OVERRIDE_GROUPS = [
-  SPLICE_AI_FIELD,
-  MOTIF_GROUP,
-  REGULATORY_GROUP,
-  SCREEN_GROUP,
-  UTR_ANNOTATOR_GROUP,
-]
-export const HIGH_MODERATE_IMPACT_GROUPS = [
-  ...HIGH_IMPACT_GROUPS,
-  VEP_GROUP_MISSENSE,
-  VEP_GROUP_INFRAME,
-]
-const CODING_IMPACT_GROUPS = [
-  VEP_GROUP_SYNONYMOUS,
-  VEP_GROUP_EXTENDED_SPLICE_SITE,
-]
-export const CODING_OTHER_IMPACT_GROUPS = [
-  ...CODING_IMPACT_GROUPS,
-  VEP_GROUP_OTHER,
-]
-
-export const ALL_ANNOTATION_FILTER = {
-  text: 'All',
-  vepGroups: ALL_IMPACT_GROUPS,
-}
-export const SV_GROUPS = [VEP_GROUP_SV_CONSEQUENCES, VEP_GROUP_SV, VEP_GROUP_SV_NEW]
-export const ANNOTATION_FILTER_OPTIONS = [
-  ALL_ANNOTATION_FILTER,
-  {
-    text: 'High Impact',
-    vepGroups: HIGH_IMPACT_GROUPS,
-  },
-  {
-    text: 'Moderate to High Impact',
-    vepGroups: HIGH_MODERATE_IMPACT_GROUPS,
-  },
-  {
-    text: 'All rare coding variants',
-    vepGroups: HIGH_MODERATE_IMPACT_GROUPS.concat(CODING_IMPACT_GROUPS),
-  },
-].map(({ vepGroups, ...option }) => ({
-  ...option,
-  value: vepGroups.reduce((acc, group) => (
-    { ...acc, [group]: GROUPED_VEP_CONSEQUENCES[group].map(({ value }) => value) }
-  ), {}),
-}))
-export const ALL_ANNOTATION_FILTER_DETAILS =
-  [ALL_ANNOTATION_FILTER].map(({ vepGroups, ...option }) => ({
-    ...option,
-    value: vepGroups.reduce((acc, group) => (
-      { ...acc, [group]: GROUPED_VEP_CONSEQUENCES[group].map(({ value }) => value) }
-    ), {}),
-  }))[0]
 
 export const THIS_CALLSET_FREQUENCY = 'callset'
 export const SV_CALLSET_FREQUENCY = 'sv_callset'
@@ -361,6 +185,10 @@ export const IN_SILICO_FIELDS = [
       }
     },
   )]
+export const IN_SILICO_SPLICING_FIELD = IN_SILICO_FIELDS.find(({ name }) => name === SPLICE_AI_FIELD)
+export const IN_SILICO_GROUP_INDEX_MAP = IN_SILICO_FIELDS.reduce(
+  (acc, { group }, i) => ({ ...acc, [group]: [...(acc[group] || []), i] }), {},
+)
 
 export const SNP_QUALITY_FILTER_FIELDS = [
   {
